+++ xd/xdfs/delimitedtext.cpp
@@ -82,7 +82,7 @@ bool DelimitedTextFile::open(const std::wstring& path, int encoding)
 {
     if (path.substr(0, 12) == L"streamptr://")
     {
-        unsigned long l = (unsigned long)kl::hexToUint64(path.substr(12));
+        uintptr_t l = (uintptr_t)kl::hexToUint64(path.substr(12));
         xd::IStream* ptr = (xd::IStream*)l;
         return open(ptr, encoding);
     }
